+++ CHANGES.rst
@@ -2,10 +2,15 @@
  change log for taxtastic
 ==========================
 
+Upcoming
+========
+
+* Updated NCBI taxdmp.zip download protocol from ftp to https [GH-156]
+
 0.10.1
 ======
 
-* bug fixes
+* Fixed broken GH-Pages build
 
 0.10.0
 ======