+++ CHANGES.rst
@@ -2,7 +2,7 @@
  change log for taxtastic
 ==========================
 
-TESTING
+TESTING2
 =======
 
 Upcoming